ejecutada version 1.1 para 08-07-2017

+++ procesaGTFS/procesaGTFS/Transporte/Grid/Grid.h
@@ -41,13 +41,13 @@ public:
 	~Grid(void);
 
 	///Coordenadas que definen grilla
-	float minLat;
-	float maxLat;
-	float minLon;
-	float maxLon;
+	double minLat;
+	double maxLat;
+	double minLon;
+	double maxLon;
 
-	float nLat;
-	float nLon;
+	double nLat;
+	double nLon;
 
 	double delta_Latitud;
 	double delta_Longitud;
